Adding extra tests to up the coverage
